+++ clang-tools-extra/clang-tidy/hooks/HooksTidyModule.cpp
@@ -13,6 +13,7 @@
 #include "AccountConvBufLenCheck.h"
 #include "AccountConvPureCheck.h"
 #include "ArrayBufLenCheck.h"
+#include "BurdenPrereqCheck.h"
 #include "EntryPointRecursionCheck.h"
 #include "EntryPointsCheck.h"
 #include "EntryPointsNegCheck.h"
@@ -45,6 +46,8 @@ public:
         "hooks-account-conv-pure");
     CheckFactories.registerCheck<ArrayBufLenCheck>(
         "hooks-array-buf-len");
+    CheckFactories.registerCheck<BurdenPrereqCheck>(
+        "hooks-burden-prereq");
     CheckFactories.registerCheck<EntryPointRecursionCheck>(
         "hooks-entry-point-recursion");
     CheckFactories.registerCheck<EntryPointsCheck>(